fix(pmon): widen the sample counters to 64 bits
count and blocks are incremented once per measured call. Multipliers are
evaluated far more often than triggers, values or actions, and with
AiPlayerbot.PerfMonEnabled left on in the config rather than toggled for
a few minutes, a busy realm reaches 2^32 increments and the report
silently wraps to a smaller number, taking avg = total / count with it.

The per-type accumulators in PrintStats wrap sooner still: they sum every
row of a metric, so they overflow roughly as many times faster as there
are names in the block.

count and blocks become uint64_t, the typeCount / typeBlocks
accumulators in both PrintStats branches and DumpJson's Sample::count
follow. No behaviour change; the fields are only ever read to format a
report.

+++ src/Bot/Debug/PerfMonitor.cpp
@@ -249,8 +249,8 @@ void PerfMonitor::PrintStats(bool perTick, bool fullStack)
             uint64 typeTotalTime = 0;
             uint64 typeMinTime = 0xffffffffu;
             uint64 typeMaxTime = 0;
-            uint32 typeCount = 0;
-            uint32 typeBlocks = 0;
+            uint64 typeCount = 0;
+            uint64 typeBlocks = 0;
             for (auto& name : names)
             {
                 PerformanceData* pd = pdMap[name];
@@ -335,8 +335,8 @@ void PerfMonitor::PrintStats(bool perTick, bool fullStack)
             uint64 typeTotalTime = 0;
             uint64 typeMinTime = 0xffffffffu;
             uint64 typeMaxTime = 0;
-            uint32 typeCount = 0;
-            uint32 typeBlocks = 0;
+            uint64 typeCount = 0;
+            uint64 typeBlocks = 0;
             for (auto& name : names)
             {
                 PerformanceData* pd = pdMap[name];
@@ -409,7 +409,7 @@ void PerfMonitor::DumpJson(bool perTick)
         uint64 totalTime;
         uint64 minTime;
         uint64 maxTime;
-        uint32 count;
+        uint64 count;
     };
 
     std::map<PerformanceMetric, std::vector<Sample>> samples;
